requirements.txt
seleniumbase~=1.49.25

pip~=20.2.3
wheel~=0.35.1
cryptography~=3.1.1
py~=1.9.0
python-dotenv~=0.14.0
setuptools~=50.3.0
pytest~=6.1.1
selenium~=3.141.0
parameterized~=0.7.4
jsonschema~=3.2.0